remove duplicate redundant code
