+++ src/query/storages/fuse/src/operations/analyze.rs
@@ -97,6 +97,7 @@ impl FuseTable {
 
                 // Generate new column statistics for snapshot
                 col_stats = reduce_block_statistics(&stats_of_columns)?;
+
                 // Status.
                 {
                     read_segment_count += chunk.len();
