+++ src/query/storages/fuse/src/operations/delete.rs
@@ -42,10 +42,10 @@ use common_sql::evaluator::BlockOperator;
 use storages_common_table_meta::meta::TableSnapshot;
 use tracing::info;
 
+use crate::operations::merge_into::CommitSink;
+use crate::operations::merge_into::TableMutationAggregator;
 use crate::operations::mutation::MutationAction;
-use crate::operations::mutation::MutationAggregator;
 use crate::operations::mutation::MutationPartInfo;
-use crate::operations::mutation::MutationSink;
 use crate::operations::mutation::MutationSource;
 use crate::operations::mutation::SerializeDataTransform;
 use crate::pipelines::Pipeline;
@@ -56,9 +56,9 @@ impl FuseTable {
     /// The flow of Pipeline is as follows:
     /// +---------------+      +-----------------------+
     /// |MutationSource1| ---> |SerializeDataTransform1|   ------
-    /// +---------------+      +-----------------------+         |      +------------------+      +------------+
-    /// |     ...       | ---> |          ...          |   ...   | ---> |MutationAggregator| ---> |MutationSink|
-    /// +---------------+      +-----------------------+         |      +------------------+      +------------+
+    /// +---------------+      +-----------------------+         |      +-----------------------+      +----------+
+    /// |     ...       | ---> |          ...          |   ...   | ---> |TableMutationAggregator| ---> |CommitSink|
+    /// +---------------+      +-----------------------+         |      +-----------------------+      +----------+
     /// |MutationSourceN| ---> |SerializeDataTransformN|   ------
     /// +---------------+      +-----------------------+
     #[async_backtrace::framed]
@@ -82,13 +82,6 @@ impl FuseTable {
         if snapshot.summary.row_count == 0 {
             // empty snapshot, no deletion
             return Ok(());
-        }
-
-        // Status.
-        {
-            let status = "delete: begin to run delete tasks";
-            ctx.set_status_info(status);
-            info!(status);
         }
 
         // check if unconditional deletion
@@ -128,8 +121,7 @@ impl FuseTable {
             return Ok(());
         }
 
-        let total_tasks = self
-            .try_add_deletion_source(ctx.clone(), &filter_expr, col_indices, &snapshot, pipeline)
+        self.try_add_deletion_source(ctx.clone(), &filter_expr, col_indices, &snapshot, pipeline)
             .await?;
         if pipeline.is_empty() {
             return Ok(());
@@ -150,24 +142,22 @@ impl FuseTable {
         pipeline.resize(1)?;
 
         pipeline.add_transform(|input, output| {
-            let aggregator = MutationAggregator::new(
+            let aggregator = TableMutationAggregator::create(
                 ctx.clone(),
+                snapshot.segments.clone(),
+                snapshot.summary.clone(),
+                self.get_block_thresholds(),
+                self.meta_location_generator().clone(),
                 self.schema(),
                 self.get_operator(),
-                self.meta_location_generator().clone(),
-                snapshot.summary.clone(),
-                snapshot.segments.clone(),
-                self.get_block_thresholds(),
-                total_tasks,
             );
             Ok(ProcessorPtr::create(AsyncAccumulatingTransformer::create(
                 input, output, aggregator,
             )))
         })?;
 
-        pipeline.add_sink(|input| {
-            MutationSink::try_create(self, ctx.clone(), snapshot.clone(), input)
-        })?;
+        pipeline
+            .add_sink(|input| CommitSink::try_create(self, ctx.clone(), snapshot.clone(), input))?;
         Ok(())
     }
 
@@ -216,7 +206,7 @@ impl FuseTable {
         col_indices: Vec<usize>,
         base_snapshot: &TableSnapshot,
         pipeline: &mut Pipeline,
-    ) -> Result<usize> {
+    ) -> Result<()> {
         let projection = Projection::Columns(col_indices.clone());
         let total_tasks = self
             .mutation_block_pruning(
@@ -227,7 +217,17 @@ impl FuseTable {
             )
             .await?;
         if total_tasks == 0 {
-            return Ok(0);
+            return Ok(());
+        }
+
+        // Status.
+        {
+            let status = format!(
+                "delete: begin to run delete tasks, total tasks: {}",
+                total_tasks
+            );
+            ctx.set_status_info(&status);
+            info!(status);
         }
 
         let block_reader = self.create_block_reader(projection, false, ctx.clone())?;
@@ -281,7 +281,7 @@ impl FuseTable {
             },
             max_threads,
         )?;
-        Ok(total_tasks)
+        Ok(())
     }
 
     #[async_backtrace::framed]
