+++ src/query/storages/fuse/src/operations/merge_into/mod.rs
@@ -18,7 +18,6 @@ pub mod mutation_meta;
 mod mutator;
 mod processors;
 
-pub use mutator::mutation_accumulator::MutationAccumulator;
 pub use processors::AppendTransform;
 pub use processors::BroadcastProcessor;
 pub use processors::CommitSink;
