+++ src/query/storages/fuse/src/operations/merge_into/mutator/mod.rs
@@ -14,4 +14,6 @@
 
 pub mod deletion_accumulator;
 pub mod merge_into_mutator;
-pub mod mutation_accumulator;
+mod mutation_accumulator;
+
+pub use mutation_accumulator::MutationAccumulator;
