+++ src/query/storages/fuse/src/operations/mutation/compact/compact_part.rs
@@ -21,7 +21,7 @@ use common_exception::ErrorCode;
 use common_exception::Result;
 use storages_common_table_meta::meta::BlockMeta;
 
-use crate::operations::merge_into::mutation_meta::mutation_log::BlockMetaIndex;
+use crate::operations::merge_into::mutation_meta::BlockMetaIndex;
 
 #[derive(serde::Serialize, serde::Deserialize, PartialEq)]
 pub struct CompactPartInfo {
