+++ src/query/storages/fuse/src/operations/replace.rs
@@ -328,6 +328,7 @@ impl FuseTable {
         // a) append TableMutationAggregator
         pipeline.add_transform(|input, output| {
             let base_segments = base_snapshot.segments.clone();
+            let base_summary = base_snapshot.summary.clone();
             let thresholds = self.get_block_thresholds();
             let location_gen = self.meta_location_generator.clone();
             let schema = self.table_info.schema();
@@ -335,6 +336,7 @@ impl FuseTable {
             let mutation_aggregator = TableMutationAggregator::create(
                 ctx.clone(),
                 base_segments,
+                base_summary,
                 thresholds,
                 location_gen,
                 schema,
